admin/requirements.txt
asgiref==3.6.0
configparser==5.3.0
dj-database-url==1.2.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ez-setup==0.9
gunicorn==20.1.0
mysql-connector-python==8.0.32
protobuf==3.20.3
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
mysqlclient
